Implement and test BreakValue; implement Statement serialization
Caught the case of an if-statement without an else-branch defined.

Statement serialization not yet tested.

Illuminated for me the fact that Archetype is not OK with attributes being created on the fly. I like that, although the silent failure is disturbing. That’s why I’d created the /K switch, though.