Implement ‘key’ correctly for Unix.
I feel funny about my implementation here, which both echoes the key pressed AND follows it with a newline.
This latter addition feels like a bug, but it does make ‘key’ like a single-character, no-return version of ‘read’.
Moreover, the animal.ach program seems to have been written to expect it.