Change '>>' operator to work like restructured-text
Consecutive '>>' tokens are converted into a single new statment
type, PARAGRAPH_OUTPUT.  Word-wrapping is handled, but empty
lines, or lines with leading white space, are considered an
end-of-paragraph indicator and left alone.

This does mean that save-files before this commit can't be
loaded by this commit and onward.

Added unit tests, and made use of the improved '>>' operator
in bare.ach (and barer.ach).