Works with GCP Cloud Run now.
Archetype is separated into a library and the `archetype_run`
executable, which is run by the Flask application.

My first choice was to use Boost Beast with the GCS C++ client
together to make a single C++ Cloud Run server, but it couldn't
be done:  bazel would not build Boost, and cmake would not build
the tree of GCP client dependencies.  Got close by using vcpkg,
but I couldn't get that to work within Docker.  In the end, this
hybrid approach was the only way across the bridge that I could
find for now.

The other approaches did work, but only on a local machine, and
with all kinds of nonsense, like a symlink for Boost, and so on.