Add hidden --inspect option to inspect an ACX file.
Generates an experimental RDF version of the basic types and
instances, as well as an index of which objects are parsed by
which words, and finally a JSON dictionary of all possible
verb and noun phrases, for potential use with other parsers.

It's a work in progress, since the output is only to standard
output right now, and the three serializations aren't
demarcated, and should be.